+++ src/main/java/com/lucasdevx/todo_list/controller/TaskController.java
@@ -16,11 +16,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.lucasdevx.todo_list.dto.TaskDTO;
+import com.lucasdevx.todo_list.exception.ObjectIsNullException;
 import com.lucasdevx.todo_list.model.Task;
 import com.lucasdevx.todo_list.service.TaskService;
 
 @RestController
-@RequestMapping("/api/task")
+@RequestMapping("/api/tasks")
 public class TaskController {
 	
 	@Autowired
@@ -28,7 +29,9 @@ public class TaskController {
 	
 	@PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public TaskDTO insert(@RequestBody TaskDTO taskDTO) {
-		
+		if(taskDTO == null) {
+			throw new ObjectIsNullException();
+		}
 		Task task = taskService.parseToEntity(taskDTO);
 		
 		return taskService.parseToDTO(taskService.insert(task));
@@ -36,6 +39,9 @@ public class TaskController {
 	
 	@GetMapping("/{id}")
 	public TaskDTO findById(@PathVariable Long id) {
+		if(id <= 0) {
+			throw new IllegalArgumentException("Invalid id");
+		}
 		return taskService.parseToDTO(taskService.findById(id));
 	}
 	
@@ -51,6 +57,18 @@ public class TaskController {
 	}
 	@PutMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public TaskDTO update(@RequestBody TaskDTO taskDTO) {
+		if(taskDTO == null) {
+			throw new ObjectIsNullException();
+		}
+		
+		if(taskDTO.id() == null) {
+			throw new NullPointerException("Id is null");
+		}
+		
+		if(taskDTO.id() <= 0) {
+			throw new IllegalArgumentException("Invalid id");
+		}
+		
 		Task task = taskService.parseToEntity(taskDTO);
 		
 		return taskService.parseToDTO(taskService.update(task));
@@ -58,6 +76,9 @@ public class TaskController {
 	
 	@DeleteMapping("/{id}")
 	public ResponseEntity<?> delete(@PathVariable Long id){
+		if(id <= 0) {
+			throw new IllegalArgumentException("Invalid id");
+		}
 		taskService.delete(id);
 		
 		return ResponseEntity.noContent().build();
